Moved benchmarking files, and made research_snapshot file
