+++ code_tests/low_cost_or_live_api_tests/test_benchmarker.py
@@ -9,13 +9,13 @@ from code_tests.unit_tests.test_forecasting.forecasting_test_manager import (
     ForecastingTestManager,
     MockBot,
 )
+from forecasting_tools.benchmarking.benchmarker import Benchmarker
 from forecasting_tools.data_models.benchmark_for_bot import BenchmarkForBot
 from forecasting_tools.data_models.questions import BinaryQuestion
 from forecasting_tools.forecast_bots.official_bots.q2_template_bot import (
     Q2TemplateBot2025,
 )
 from forecasting_tools.forecast_bots.template_bot import TemplateBot
-from forecasting_tools.forecast_helpers.benchmarker import Benchmarker
 
 logger = logging.getLogger(__name__)
