+++ forecasting_tools/__init__.py
@@ -50,6 +50,12 @@ from forecasting_tools.ai_models.general_llm import GeneralLlm as GeneralLlm
 from forecasting_tools.ai_models.resource_managers.monetary_cost_manager import (
     MonetaryCostManager as MonetaryCostManager,
 )
+from forecasting_tools.benchmarking.benchmark_displayer import (
+    run_benchmark_streamlit_page as run_benchmark_streamlit_page,
+)
+from forecasting_tools.benchmarking.benchmarker import (
+    Benchmarker as Benchmarker,
+)
 from forecasting_tools.data_models.benchmark_for_bot import (
     BenchmarkForBot as BenchmarkForBot,
 )
@@ -120,12 +126,6 @@ from forecasting_tools.forecast_bots.template_bot import (
 )
 from forecasting_tools.forecast_helpers.asknews_searcher import (
     AskNewsSearcher as AskNewsSearcher,
-)
-from forecasting_tools.forecast_helpers.benchmark_displayer import (
-    run_benchmark_streamlit_page as run_benchmark_streamlit_page,
-)
-from forecasting_tools.forecast_helpers.benchmarker import (
-    Benchmarker as Benchmarker,
 )
 from forecasting_tools.forecast_helpers.metaculus_api import (
     ApiFilter as ApiFilter,
